Move services to separate directory
