+++ services/shortcut-service.js
